requirements.txt
aiohttp
beautifulsoup4
bs4
faiss-cpu
librosa
numpy
pandas
protobuf
pinecone-client
torch
transformers
tf-keras
openai-whisper
scipy
soundfile
langchain
langchain-huggingface
langchain-google-genai
langchain-community
sentence-transformers
flask
flask_cors
chardet
trafilatura
flashrank
watchdog

